requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
lxml==4.9.3
pycryptodomex==3.18.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
steam==1.4.4
urllib3==2.0.4
vdf==3.4
